Stops trying to use the objectmapper in Scala client
The combination of vertx's Json object plus Jackson-databind plus Scala turns out to be ugly.

This really leaves a choice between switching to a more scala-friendly JSON library (e.g. Circe) or just writing mappings from Vertx JsonObjects and JsonArrays manually

As I don't teach Circe anywhere, I think that'd be an awkward learning burden for students, so in this commit I switch the code to doing it the longwinded-but-easier-to-understand way: each of the data classes simply has a "fromJson" method defined that will take a vertx jsonObject and manually convert it into an instance of the data class.